refactor: convert crud package to a template
